ref: Layoutsを廃止して各コンポーネントで読み込み
タイトル表示を流し込みたいため
